Only download one track at a time (fixes #14)
It appears that starting all downloads for an album increases the risk
of corrupting one of the tracks, especially if one of those tracks is
very long.  This change chains the promises for downloaded the tracks
sequentially instead of parrallelly.

Given that the bottleneck for downloading many files is
network bandwidth, downloading all the files in paralell should not
increase the wall-time of the download.

Unforunately in practice, this does not prove to be the case.  Emperical
tests have shown this to be about 50% slower, so this may not be the
best approach.